+++ resources/views/carratage/index.blade.php
@@ -124,6 +124,7 @@
                     order: [
                         [0, 'desc']
                     ],
+                    autoWidth: false,
                     processing: true,
                     serverSide: true,
                     ajax: '{{ route('fetchcarrat') }}',
@@ -175,6 +176,12 @@
                         _token: _token
                     },
                     success: function(data) {
+                        const Toast = Swal.mixin({
+                            toast: true,
+                            position: 'top-end',
+                            showConfirmButton: false,
+                            timer: 3000
+                        });
                         if (data.status == 400) {
                             $('#saveform_errList').html("");
                             $('#saveform_errList').addClass("alert alert-danger");
@@ -183,10 +190,13 @@
                             });
                         } else {
                             $("#saveform_errList").html("");
-                            $("#success_message").addClass('alert alert-success')
-                            $("#success_message").text(data.message)
                             $("#addTodoModal").modal('hide');
                             $("#addTodoModal").find('input').val("");
+                            Toast.fire({
+                                type: 'success',
+                                title: data.message,
+                            });
+
                             // $("#addTodoModal").find('textarea').val("");
 
                             $("#data-table").dataTable().fnDestroy();
@@ -208,20 +218,49 @@
             function deleteTodo(id) {
                 let url = `carrat/${id}`;
                 let token = $('meta[name="csrf-token"]').attr('content');
+                const Toast = Swal.mixin({
+                    toast: true,
+                    position: 'top-end',
+                    showConfirmButton: false,
+                    timer: 3000
+                });
 
-                $.ajax({
-                    url: url,
-                    type: 'DELETE',
-                    data: {
-                        id: id,
-                        _token: token
-                    },
-                    success: function(data) {
-                        $("#todo_" + data.data.id).remove();
+                Swal.fire({
+                    title: "Are you sure?",
+                    text: "You will not be able to recover this imaginary file!",
+                    type: "warning",
+                    showCancelButton: true,
+                    confirmButtonColor: '#DD6B55',
+                    confirmButtonText: 'Yes, delete it!',
+                    closeOnConfirm: true
+                }).then((result) => {
+                    /* Read more about isConfirmed, isDenied below */
+                    if (result) {
+
+                        $.ajax({
+                            type: 'DELETE',
+                            headers: {
+                                'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
+                            },
+                            url: url,
+                            data: {
+                                _token: token
+                            },
+                            success: function(response, textStatus, xhr) {
+                                Toast.fire({
+                                    type: 'success',
+                                    title: 'Carratage Deleted Successfully!',
+                                    customClass: {
+                                        popup: 'adjust'
+                                    }
+                                })
+                                $("#data-table").dataTable().fnDestroy();
+                                fetchCarrat();
+                            }
+                        });
                     }
                 });
-                $("#data-table").dataTable().fnDestroy();
-                fetchCarrat();
+
             }
 
             function editTodo(id) {
@@ -249,6 +288,12 @@
                 var id = $('#edit_todo_id').val();
                 let _url = `/carrat/${id}`;
                 let _token = $('meta[name="csrf-token"]').attr('content');
+                const Toast = Swal.mixin({
+                            toast: true,
+                            position: 'top-end',
+                            showConfirmButton: false,
+                            timer: 3000
+                        });
 
                 $.ajax({
                     url: _url,
@@ -267,14 +312,19 @@
                             });
                         } else {
                             $("#edit_saveform_errList").html("");
-                            $("#success_message").addClass('alert alert-success')
-                            $("#success_message").text(data.message)
                             $("#addTodoModal").modal('hide');
                             $("#addTodoModal").find('input').val("");
                             $('#edit_todo_id').val('');
                             $('#edit_carrat_value').val('');
                             $('#edit_val_per_g').val('');
                             $('#editTodoModal').modal('hide');
+                            Toast.fire({
+                                    type: 'success',
+                                    title: data.message,
+                                    customClass: {
+                                        popup: 'adjust'
+                                    }
+                                })
                             $("#data-table").dataTable().fnDestroy();
                             fetchCarrat();
                         }
